api_automation_using_pyton/requirements.txt
certifi==2024.12.14
charset-normalizer==2.1.1
exceptiongroup==1.3.0
idna==3.3
iniconfig==1.1.1
packaging==24.2
pluggy==1.0.0
pytest==7.1.2
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.12
pydantic
python-dotenv==1.0.0
